split pth to be able to run standalone and in website
